Removed * in searchFilter used for getting e-mail from LDAP Apps Support ticket was submitted (31169 - incorrect e-mail associated with my caMOD account) stating:
Cheryl Marks submitted a ticket to NCICB Application Support in which she requested that the e-mail address associated with her account in the "User Settings" screen in caMOD be corrected. She has attempted to correct it herself, but because the program queries the LDAP Server for the e-mail address, her corrections were not retained.


SVN-Revision: 3328