modified for gfrorge #16902  	link to the induced mutation page in the Admin/ Edited Models mode is not available for old models
Reversed last two code fixes in subSubmitMenu.jsp.  This fix needs to be done in this class.  The EnvironmentalFactor object has a convinience method to get the name we want to display.  The getDisplayNameIM method was modified to look for nameAlternEntry or name (for older models) when displaying IM models.

We have decided to rework this since we have the EF flag after release 2.5 and nameAlternEntry no longer may be needed to differeentiate between CE and IM in the EF table.


SVN-Revision: 3923