modified for gforge #17320  	Make sure all field sizes in database are validated correctly in validation.xml
The validation is triggered when I open the image page of an existing record and want to resubmit the page. The validation is asking for an image or an entry in the alternative URL field. Validation should recognize when there is data in the db for these two fields.

Validation was taken out.  The validwhen or requiredif rules can not cover all cases.


SVN-Revision: 4338